Fix argument type error: cast monthlyDelta to double for _formatConsumptionDelta

+++ lib/screens/post_sign_in_page.dart
@@ -609,9 +609,9 @@ class _PostSignInPageState extends State<PostSignInPage>
                     width: math.max(metricWidth, 150),
                     child: _consumptionMetricCard(
                       label: _t('Month change', 'የወር ለውጥ'),
-                      value: currentPoint == null || previousPoint == null
+                        value: currentPoint == null || previousPoint == null
                           ? _t('No comparison', 'ንፅፅር የለም')
-                          : _formatConsumptionDelta(monthlyDelta),
+                          : _formatConsumptionDelta(monthlyDelta.toDouble()),
                       subtext: currentPoint == null || previousPoint == null
                           ? _t(
                               'Need at least two months.',
